add test snapshot for win32.
