+++ test/snapshots/win32/counterTest/counterTest.json
@@ -1,0 +1,1 @@
+{"D:\\test\\fixture\\counterTest\\counterTest.vue":1,"D:\\test\\components\\nested2.vue":1,"D:\\test\\components\\child.vue":2,"D:\\test\\components\\child2.vue":2,"D:\\test\\fixture\\counterTest\\counterComponent.vue":1,"D:\\test\\fixture\\counterTest\\counterComponent2.vue":1}
