+++ test/snapshots/win32/declarationTest/SJIS.json
@@ -1,0 +1,1 @@
+{"name":"\\test\\fixture\\declarationTest\\SJIS.vue","props":"","size":103,"lastModifiedTime":0,"children":[{"name":"\\test\\components\\child.vue","props":"","size":42,"lastModifiedTime":0,"children":[]}]}
