+++ test/snapshots/win32/declarationTest/importNestedOne.json
@@ -1,0 +1,1 @@
+{"name":"\\test\\fixture\\declarationTest\\importNestedOne.vue","props":"","size":108,"lastModifiedTime":0,"children":[{"name":"\\test\\components\\nested.vue","props":"","size":97,"lastModifiedTime":0,"children":[{"name":"\\test\\components\\child.vue","props":"","size":42,"lastModifiedTime":0,"children":[]}]}]}
