+++ test/snapshots/win32/declarationTest/importNestedTwo.json
@@ -1,0 +1,1 @@
+{"name":"\\test\\fixture\\declarationTest\\importNestedTwo.vue","props":"","size":162,"lastModifiedTime":0,"children":[{"name":"\\test\\components\\nested.vue","props":"","size":97,"lastModifiedTime":0,"children":[{"name":"\\test\\components\\child.vue","props":"","size":42,"lastModifiedTime":0,"children":[]}]},{"name":"\\test\\components\\nested2.vue","props":"","size":135,"lastModifiedTime":0,"children":[{"name":"\\test\\components\\child.vue","props":"","size":42,"lastModifiedTime":0,"children":[]},{"name":"\\test\\components\\child2.vue","props":"","size":43,"lastModifiedTime":0,"children":[]}]}]}
