+++ test/snapshots/win32/declarationTest/importTwo.json
@@ -1,0 +1,1 @@
+{"name":"\\test\\fixture\\declarationTest\\importTwo.vue","props":"","size":158,"lastModifiedTime":0,"children":[{"name":"\\test\\components\\child.vue","props":"","size":42,"lastModifiedTime":0,"children":[]},{"name":"\\test\\components\\child2.vue","props":"","size":43,"lastModifiedTime":0,"children":[]}]}
