+++ test/snapshots/win32/declarationTest/noExtension.json
@@ -1,0 +1,1 @@
+{"name":"\\test\\fixture\\declarationTest\\noExtension.vue","props":"","size":147,"lastModifiedTime":0,"children":[{"name":"\\test\\components\\nested.vue","props":"","size":97,"lastModifiedTime":0,"children":[{"name":"\\test\\components\\child.vue","props":"","size":42,"lastModifiedTime":0,"children":[]}]},{"name":"\\test\\assets\\module","props":"","size":0,"lastModifiedTime":0,"children":[]}]}
