+++ test/snapshots/win32/declarationTest/onlyTemplate.json
@@ -1,0 +1,1 @@
+{"name":"\\test\\fixture\\declarationTest\\onlyTemplate.vue","props":"","size":37,"lastModifiedTime":0,"children":[]}
