+++ test/snapshots/win32/propsTest/oneProps.json
@@ -1,0 +1,1 @@
+{"name":"\\test\\fixture\\propsTest\\oneProps.vue","props":{"data":{"type":"String","required":"true","default":"1"}},"size":192,"lastModifiedTime":0,"children":[]}
