+++ test/specs/props.test.ts
@@ -1,16 +1,19 @@
-const {getImportDeclarationTree} = require('../../src/utils');
+import {platform} from 'os';
 import {existsSync, writeFileSync, readFileSync} from 'fs';
 import {join, dirname} from 'path';
 import assert from 'assert';
 const mkdirp = require('mkdirp');
+const {Analyzer} = require('../../src/server/Analyzer');
 const fixturesDir = join(__dirname, '../fixture/');
-const snapshotDir = join(__dirname, '../snapshots/');
+const snapshotDir = platform() === 'win32' ? join(__dirname, '../snapshots/win32') : join(__dirname, '../snapshots/');
 const isUpdate = process.argv.includes('--update');
+
+const analyzer = new Analyzer();
 
 describe('props test', () => {
   it('oneProps.vue', async () => {
     const filename = 'propsTest/oneProps';
-    const declaration = await getImportDeclarationTree(join(fixturesDir, `${filename}.vue`), true);
+    const declaration = await analyzer.getImportDeclarationTree(join(fixturesDir, `${filename}.vue`), true);
     const expectedFilePath = join(snapshotDir, `${filename}.json`);
     const json = JSON.stringify(declaration).slice(0);
 
@@ -31,7 +34,7 @@ describe('props test', () => {
 
   it('twoProps.vue', async () => {
     const filename = 'propsTest/twoProps';
-    const declaration = await getImportDeclarationTree(join(fixturesDir, `${filename}.vue`), true);
+    const declaration = await analyzer.getImportDeclarationTree(join(fixturesDir, `${filename}.vue`), true);
     const expectedFilePath = join(snapshotDir, `${filename}.json`);
     const json = JSON.stringify(declaration).slice(0);
